app/requirements.txt
anyio==3.6.2
autopep8==2.0.1
click==8.1.3
colorama==0.4.6
fastapi==0.89.1
greenlet==2.0.1
h11==0.14.0
idna==3.4
pycodestyle==2.10.0
pydantic==1.10.4
PyMySQL==1.0.2
python-dotenv==0.21.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
